Example 1
This is an example of the schedule format that I created the code to pull from. The parts in red represent the name of the opponent's school. Any opponent with an “@“ in front represents an away game, and any opponent without one represents a home game (see Women’s Soccer Table). A limitation to my program is that it only works for one sport’s section at a time. This is an actual schedule that the Athletic Director used, except I made a copy and modified some dates to make the test run easier. I blocked off any potentially revealing information due to the fact that I didn’t ask for permission to share this.